requirements.txt
RPi.GPIO==0.7.1a4
adafruit_dht==1.4.0
psycopg2-binary==2.9.9
python-dotenv==0.21.0